backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
python-multipart==0.0.12
google-generativeai==0.8.3
aiofiles==24.1.0
python-dotenv==1.0.1
